requirements.txt
tavily-python==0.5.1
pymilvus==2.5.6
numpy==2.2.4
tenacity==9.0.0
python-dotenv==1.0.1
langchain-core==0.3.51
langchain-community==0.3.17
langchain-milvus==0.1.9
langchain-openai==0.3.12
langchain-deepseek==0.1.3
langchain-huggingface==0.1.2
huggingface_hub[hf_xet]==0.30.2
langgraph-cli[inmem]